# Optimize NumberBufferToDecimal