This is the last managed decimal function slower than the previous native equivalent.
Perf stats for `decimal.TryParse(s, NumberStyles.Number, CultureInfo.InvariantCulture, out _)`:

### x64
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 175.3 ns | 0.3196 ns | 0.0830 ns |   1.00 |
| CoreRT | 259.5 ns | 0.7716 ns | 0.2752 ns |   1.48 |
| CoreRT2 | 113.8 ns | 0.3476 ns | 0.1240 ns |   0.65 |

### x86
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 151.0 ns | 0.2238 ns | 0.0798 ns |   1.00 |
| CoreRT | 255.0 ns | 0.6094 ns | 0.2173 ns |   1.69 |
| CoreRT2 | 127.5 ns | 0.5105 ns | 0.1820 ns |   0.84 |