@jkotas 2018-08-14
@pentp Could you please incorporate the feedback so that this can be merged? Thanks!